Added new Edit-function

+++ src/main/java/de/santax/dashboard/teacher/TeacherService.java
@@ -1,10 +1,12 @@
 package de.santax.dashboard.teacher;
 
 import de.santax.dashboard.student.exception.BadRequestException;
+import de.santax.dashboard.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -16,16 +18,35 @@ public class TeacherService {
         return teacherRepository.findAll();
     }
 
+    public Teacher getTeacherById(Long id) {
+        return teacherRepository.findById(id).orElseThrow(() ->
+                new BadRequestException("Cannot find Teacher with ID: " + id));
+    }
+
     public void addTeacher(Teacher teacher) {
         boolean existingTeacher = teacherRepository.existingEmail(teacher.getEmail());
-        if(existingTeacher) {
+        if (existingTeacher) {
             throw new BadRequestException("Email " + teacher.getEmail() + " is already taken");
         }
         teacherRepository.save(teacher);
     }
 
     public void deleteTeacher(Long id) {
+        Optional<Teacher> existingTeacher = teacherRepository.findById(id);
+        if (existingTeacher.isEmpty()) {
+            throw new StudentNotFoundException("Teacher with ID: " + id + " not found.");
+        }
         teacherRepository.deleteById(id);
     }
 
+    public void updateTeacher(Teacher teacher) {
+        Teacher updatedTeacher = teacherRepository.findById(teacher.getId()).orElseThrow(() ->
+                new BadRequestException("Cannot update Teacher" + teacher.getName())
+        );
+        updatedTeacher.setName(teacher.getName());
+        updatedTeacher.setEmail(teacher.getEmail());
+        updatedTeacher.setGender(teacher.getGender());
+
+        teacherRepository.save(updatedTeacher);
+    }
 }
